fix: Correct sign compare warning message

+++ Sources/Rosetta/Actions/PlayCard.cpp
@@ -121,7 +121,8 @@ bool IsPlayableByCardReq(Entity* source)
             case PlayReq::REQ_MINIMUM_ENEMY_MINIONS:
             {
                 auto& opField = source->GetOwner().GetOpponent().GetField();
-                if (opField.GetNumOfMinions() < requirement.second)
+                if (static_cast<int>(opField.GetNumOfMinions()) <
+                    requirement.second)
                 {
                     return false;
                 }
